Fixed library corruption due to bad file extension filtering

+++ Resources/build-sgl302.sh
@@ -20,13 +20,12 @@ if [ $INSTALL_SGL_LIB -eq 1 ]; then
 	#
 	# Create destination directories
 	#
-	mkdir -p $1/{lib,include,samples/{sample1,sample2},demos,doc,sddrv}
+	mkdir -p $1/{common,lib,include,samples/{sample1,sample2},demos,doc,sddrv}
 
 	#
 	# convert coff to ELF
 	#
 	find $SATURN_TMP/sgl302/lib/ \
-	 ! -name "libsgl_real_elf.a" \
 	 -name "*.[a,o]" \
 	 -type f \
 	 -exec "$SATURN_ROOT/toolchain/bin/${PROGRAM_PREFIX}objcopy" -v -Icoff-sh -Oelf32-sh {} \;
@@ -34,9 +33,15 @@ if [ $INSTALL_SGL_LIB -eq 1 ]; then
 	#
  	# Install libsgl patched version
  	#
-	if [ -f "$SATURN_TMP/sgl302/lib/libsgl_real_elf.a" ]; then
-		mv -fv "$SATURN_TMP/sgl302/lib/libsgl_real_elf.a" "$SATURN_TMP/sgl302/lib/libsgl.a"
+	if [ -f "$SATURN_TMP/sgl302/libsgl_real_elf.a" ]; then
+		mv -fv "$SATURN_TMP/sgl302/libsgl_real_elf.a" "$SATURN_TMP/sgl302/lib/libsgl.a"
 	fi
+
+	# Clean the code
+	find $SATURN_TMP/sgl302 \
+		! -name "*.[a,o,lib]" \
+		-type f \
+		-exec sed -i 's/\o32//g' {} \;
 
 	#
 	# Copy to destination
@@ -49,11 +54,10 @@ if [ $INSTALL_SGL_LIB -eq 1 ]; then
 	cp -rv $SATURN_TMP/sgl302/doc/* $1/doc
 	cp -rv $SATURN_TMP/sgl302/sddrv/* $1/sddrv
 
-	# Removed, will come from SBL
-	rm $1/include/sl_def.h
-
-	# Clean the code
-	find $1 -type f -exec sed -i 's/\o32//g' {} \;
+	if [ $INSTALL_SBL_LIB -eq 0 ]; then
+		# Removed, will come from SBL
+		rm $1/include/sl_def.h
+	fi
 
 else
 	echo "$(tput setaf 1)No SGL libraries will be built$(tput sgr 0)"
